move libinfo async-start test into resolve-dns.test.ts; filter ASAN stderr
Per review: keep the new test alongside the existing dns.lookup coverage
instead of a standalone file, and strip the 'WARNING: ASAN interferes'
startup line before asserting stderr is empty so debug/ASAN builds on
macOS do not spuriously fail.
